port restaurantbot to sdk
