+++ examples/restaurantbot/Makefile
@@ -7,7 +7,14 @@ help:
 	@echo "        Runs the bot on the command line."
 
 run:
-	python bot.py run
+	make run-actions&
+	make run-core
+
+run-actions:
+	python -m rasa_core_sdk.endpoint --actions actions
+
+run-core:
+	python -m rasa_core.run --nlu models/nlu/default/current --core models/dialogue --endpoints endpoints.yml
 
 train-nlu:
 	python bot.py train-nlu
