+++ examples/restaurantbot/bot.py
@@ -9,14 +9,7 @@ import warnings
 
 from policy import RestaurantPolicy
 from rasa_core import utils
-from rasa_core.actions import Action
 from rasa_core.agent import Agent
-from rasa_core.channels.console import ConsoleInputChannel
-from rasa_core.events import SlotSet
-from rasa_core.featurizers import (
-    MaxHistoryTrackerFeaturizer,
-    BinarySingleStateFeaturizer)
-from rasa_core.interpreter import RasaNLUInterpreter
 from rasa_core.policies.memoization import MemoizationPolicy
 
 logger = logging.getLogger(__name__)
@@ -25,30 +18,6 @@ logger = logging.getLogger(__name__)
 class RestaurantAPI(object):
     def search(self, info):
         return "papi's pizza place"
-
-
-class ActionSearchRestaurants(Action):
-    def name(self):
-        return 'action_search_restaurants'
-
-    def run(self, dispatcher, tracker, domain):
-        dispatcher.utter_message("looking for restaurants")
-        restaurant_api = RestaurantAPI()
-        restaurants = restaurant_api.search(tracker.get_slot("cuisine"))
-        return [SlotSet("matches", restaurants)]
-
-
-class ActionSuggest(Action):
-    def name(self):
-        return 'action_suggest'
-
-    def run(self, dispatcher, tracker, domain):
-        dispatcher.utter_message("here's what I found:")
-        dispatcher.utter_message(tracker.get_slot("matches"))
-        dispatcher.utter_message("is it ok for you? "
-                                 "hint: I'm not going to "
-                                 "find anything else :)")
-        return []
 
 
 def train_dialogue(domain_file="restaurant_domain.yml",
@@ -84,15 +53,6 @@ def train_nlu():
     return model_directory
 
 
-def run(serve_forever=True):
-    interpreter = RasaNLUInterpreter("models/nlu/default/current")
-    agent = Agent.load("models/dialogue", interpreter=interpreter)
-
-    if serve_forever:
-        agent.handle_channel(ConsoleInputChannel())
-    return agent
-
-
 if __name__ == '__main__':
     utils.configure_colored_logging(loglevel="INFO")
 
@@ -110,5 +70,3 @@ if __name__ == '__main__':
         train_nlu()
     elif task == "train-dialogue":
         train_dialogue()
-    elif task == "run":
-        run()
